refactor: reorganize code structure and rename packages
