+++ app/src/androidTest/java/com/github/orator/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.project
+package com.github.orator
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
