+++ app/src/main/java/com/github/se/orator/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.project
+package com.github.se.orator
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -11,7 +11,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.project.ui.theme.ProjectTheme
+import com.github.se.orator.ui.theme.ProjectTheme
 
 class MainActivity : ComponentActivity() {
   override fun onCreate(savedInstanceState: Bundle?) {
