+++ app/src/main/java/com/github/se/orator/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.project.ui.theme
+package com.github.se.orator.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
